refactored component structure
