+++ pages/rank/[category].jsx
@@ -5,7 +5,7 @@ import cookies from 'next-cookies';
 import parseCookies from 'utils/parseCookies';
 
 import Header from 'containers/Header';
-import RankGame from 'containers/RankGame';
+import RankGame from 'containers/RankGame/Game';
 import Four0Four from 'containers/Four0Four';
 
 /**
